Add Info Panel to direct chats
