+++ src/screens/Messages/components/MessagesList.tsx
@@ -65,6 +65,7 @@ import {ChatStatusInfo} from './ChatStatusInfo'
 import {groupSystemMessages, type RenderItem} from './groupSystemMessages'
 import {InviteLinkDialogProvider} from './InviteLinkDialogProvider'
 import {MessageInputEmbed, useMessageEmbed} from './MessageInputEmbed'
+import {MessagesListGroupInfoPanel} from './MessagesListGroupInfoPanel'
 import {MessagesListInfoPanel} from './MessagesListInfoPanel'
 import {KeyboardStickyView} from './vendor/KeyboardStickyView'
 
@@ -509,9 +510,12 @@ export function MessagesList({
             ListHeaderComponent={
               <>
                 <MaybeLoader isLoading={convoState.isFetchingHistory} />
-                {convoState.convo?.kind === 'group' &&
-                convoState.hasAllHistory ? (
-                  <MessagesListInfoPanel convo={convoState.convo} />
+                {convoState.hasAllHistory ? (
+                  convoState.convo?.kind === 'group' ? (
+                    <MessagesListGroupInfoPanel convo={convoState.convo} />
+                  ) : (
+                    <MessagesListInfoPanel convo={convoState.convo} />
+                  )
                 ) : null}
               </>
             }
